Added a custom ordering feature for notes which does so by the following
Adding an extra column for every note called 'allNotesPosition' on the backend, the front-end uses it to store the order of the notes using a method much like inserting a note between orders 1 and 2 with an order 1.5 but instead using visible values on the ASCII table to make a base-95 system to store it as a string.

This method allows SO MUCH better precision in a compressed amount of space with no capacity limit on the number of divisions, though it's values are not as easily compared as numbers, which has made it harder to render them in O(n) time.

My current method is to pre-process a userOrder array, which has the potential to be as efficient as any sorted list using binary insert and search